@@ -7,30 +7,72 @@
   "private": true,
   "scripts": {
     "compile": "hardhat compile",
-    "test": "hardhat test"
+    "test": "hardhat test",
+    "clean": "node utils/clean.js",
+    "prettier:contracts": "prettier --write contracts/**/*.sol",
+    "prettier:tests": "prettier --write test/**/*.ts",
+    "goRinkeby": "hardhat run scripts/auction.js --network rinkeby",
+    "payments-mainnet": "hardhat run scripts/payments.js --network mainnet",
+    "buy": "hardhat run scripts/buy-pill.js --network mainnet",
+    "limit": "hardhat run scripts/set-limit.js --network mainnet",
+    "collect": "hardhat run scripts/collect.js --network mainnet",
+    "char-mock-rink": "hardhat run scripts/char-mockup.ts --network rinkeby",
+    "minters-txt": "hardhat run scripts/get-all-minters.ts --network rinkeby",
+    "char-mock-info": "hardhat run scripts/get-char-mock-info.ts --network rinkeby",
+    "goMainnet": "hardhat run scripts/auction.js --network mainnet",
+    "deployRinkeby": "hardhat run scripts/deploy.js --network rinkeby",
+    "characters:arb": "hardhat run scripts/characters.ts --network arbrinkeby",
+    "characters:mint:arb": "hardhat run scripts/mint-character.ts --network arbrinkeby",
+    "deployMainnet": "hardhat run scripts/deploy.js --network mainnet",
+    "airtable": "ts-node-esm ./scripts/airtable.ts"
+
   },
   "devDependencies": {
+    "@nomiclabs/hardhat-ethers": "^2.0.5",
+    "@nomiclabs/hardhat-etherscan": "^3.0.3",
     "@nomiclabs/hardhat-solhint": "^2.0.0",
-    "@nomiclabs/hardhat-truffle5": "^2.0.0",
+    "@nomiclabs/hardhat-waffle": "^2.0.3",
     "@nomiclabs/hardhat-web3": "^2.0.0",
-    "@openzeppelin/contracts": "^4.0.0",
+    "@openzeppelin/contracts": "^4.6.0",
+    "@chainlink/contracts": "0.4.1",
     "@openzeppelin/test-helpers": "^0.5.10",
-    "@truffle/contract": "^4.3.13",
-    "@types/chai": "^4.2.16",
-    "@types/mocha": "^8.2.2",
-    "@types/node": "^14.14.41",
-    "chai": "^4.3.3",
-    "hardhat": "^2.2.0",
-    "hardhat-gas-reporter": "^1.0.4",
+    "@typechain/ethers-v5": "^10.0.0",
+    "@typechain/hardhat": "^6.0.0",
+    "@types/chai": "^4.3.1",
+    "@types/mocha": "^9.1.0",
+    "@types/node": "^17.0.25",
+    "chai": "^4.3.6",
+    "dotenv": "^16.0.0",
+    "ethereum-waffle": "^3.4.4",
+    "hardhat-gas-reporter": "^1.0.8",
     "hardhat-watcher": "^2.1.1",
-    "mocha": "^8.3.0",
-    "prettier-plugin-solidity": "^1.0.0-beta.5",
+    "mocha": "^9.2.2",
+    "prettier": "^2.6.2",
+    "prettier-plugin-solidity": "^1.0.0-beta.19",
+    "pretty-quick": "^3.1.3",
     "solhint": "^3.3.4",
-    "solidity-coverage": "^0.7.16",
-    "truffle": "^5.3.2",
-    "truffle-hdwallet-provider": "^1.0.17",
-    "ts-node": "^9.1.1",
-    "typescript": "^4.2.4",
-    "web3": "^1.3.5"
+    "solidity-coverage": "^0.7.20",
+    "ts-node": "^10.8.0",
+    "typechain": "^8.0.0",
+    "typescript": "^4.6.3",
+    "web3": "^1.3.5",
+    "arb-bridge-eth": "0.7.7",
+    "@types/lodash": "^4.14.182",
+    "type-fest": "^2.14.0"
+  },
+  "dependencies": {
+    "@ethersproject/abi": "^5.6.1",
+    "@ethersproject/bytes": "^5.6.1",
+    "@ethersproject/providers": "^5.6.4",
+    "@rari-capital/solmate": "^6.2.0",
+    "@textile/hub": "^6.3.4",
+    "airtable": "^0.11.4",
+    "cross-env": "^7.0.3",
+    "ethers": "^5.6.4",
+    "hardhat": "^2.9.3",
+    "hardhat-etherscan": "^1.0.1",
+    "lodash": "^4.17.21",
+    "merge-images": "^2.0.0",
+    "ts-extras": "^0.10.2"
   }
 }
