@@ -1,47 +1,42 @@
-/* global artifacts, it, assert */
-/* eslint-disable prefer-reflect */
-const { artifacts, web3, contract } = require("hardhat");
-import { assert } from "chai";
-const { expectEvent, expectRevert } = require("@openzeppelin/test-helpers");
+import { ethers } from "hardhat";
+import { expect, assert } from "chai";
+import { MetadataRegistry } from "../typechain-types/contracts/MetadataRegistry.sol";
+import { SignerWithAddress } from "@nomiclabs/hardhat-ethers/signers";
+import { deployRegistry } from "./shared/deploys";
 
-const Registry = artifacts.require("MetadataRegistry.sol");
+describe("MetadataRegistry", async () => {
+  let user2: SignerWithAddress, user1: SignerWithAddress;
+  const registry = (await deployRegistry()) as MetadataRegistry;
+  [user1, user2] = await ethers.getSigners();
 
-const toBN = (x: string | number) => web3.utils.toBN(x);
-
-contract("MetadataRegistry", (addresses) => {
-  const user1 = addresses[0];
-  let registry;
   const metadata = "hello world";
-  const id = toBN(0);
-  const badId = toBN(1);
-  it("deploys", async () => {
-    registry = await Registry.new();
-  });
+  const id = 0;
+  const badId = 1;
+
   it("can set metadata", async () => {
-    const receipt = await registry.set(0, "hello world");
-    expectEvent(receipt, "Register", { id, metadata });
+    await expect(registry.set(0, "hello world"))
+      .to.emit(registry, "Register")
+      .withArgs(id, metadata);
   });
   it("can get metadata", async () => {
     const md = await registry.get(id);
     assert.equal(md, metadata);
   });
   it("reverts if metadata doesn't exist", async () => {
-    await expectRevert(registry.get(badId), "MISSING_URI");
+    expect(registry.get(badId)).to.be.revertedWith("MISSING_URI");
   });
   it("only allows owner to set metadata", async () => {
-    await expectRevert(
-      registry.set(id, "hello Mars", { from: user1 }),
+    expect(registry.connect(user2).set(id, "hello Mars")).to.be.revertedWith(
       "Ownable: caller is not the owner"
     );
   });
   describe("setMultiple", () => {
     it("accepts multiple entries", async () => {
-      const keys = [1, 2, 3, 4].map(toBN);
+      const keys = [1, 2, 3, 4];
       const values = ["Hi", "mom", "from", "space"];
-      const receipt = await registry.setMultiple(keys, values);
+      await registry.setMultiple(keys, values);
       keys.forEach(async (k, i) => {
-        assert.equal(await registry.get(k), values[i]);
-        expectEvent(receipt, "Register", { id: k, metadata: values[i] });
+        expect(await registry.get(k)).to.equal(values[i]);
       });
     });
   });
